Swagger config

+++ src/controllers/ToolsController.js
@@ -1,46 +1,47 @@
 const Tools = require('../models/Tools')
+const mongoose = require('mongoose')
 
 module.exports = {
 
-    async findAll(req, res){
+    async findAll(req, res) {
 
-        console.log("Printando a query " + req.query)
-        if(req.query.tag !== undefined){
-            console.log(" *** Query ***")
+        let tools = null
+        if (req.query.tag !== undefined) {
             const { tag } = req.query
-            console.log("Tag " + tag)
-            const tools = await Tools.find({ tags: tag })
-            return res.json(tools)
+            tools = await Tools.find({ tags: tag })
         } else {
-            console.log(" *** Sem Query ***")
-            const tools = await Tools.find().sort('-createdAt')
-            tools.map(t => console.log(t.title))
-            return res.json(tools)
+            tools = await Tools.find().sort('-createdAt')
+            //tools.map(t => console.log(t.title))            
         }
 
-        
-        return res.send()
+        if (!tools) return res.status(404).json({ error: "Dados nao encontrados" })
 
-        // await Tools.find().sort('-createdAt')
-        //     .then(tools => {
-        //         //console.log("Promise " + tools)
-        //         res.json(tools)
-        //     })
-        //     .catch(err => err)
+        return res.json(tools)
+
 
     },
 
-    async findById(req,res){
+    async findById(req, res) {
         const { id } = req.params
-        if(!id) return res.status(400).json({ error: "ID nao informado"})
+
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            return res.status(404).json({ error: "ID invalido" })
+        }
+
+        if (!id) return res.status(400).json({ error: "ID nao informado" })
 
         const tool = await Tools.findById(id)
         return res.json(tool)
     },
 
-    async update(req,res){
+    async update(req, res) {
         const { id } = req.params
-        if(!id) return res.status(400).json({ error: "ID nao informado"})
+
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            return res.status(404).json({ error: "ID invalido" })
+        }
+
+        if (!id) return res.status(400).json({ error: "ID nao informado" })
 
         const tool = await Tools.findByIdAndUpdate(id, req.body, {
             new: true
@@ -48,22 +49,31 @@ module.exports = {
         return res.json(tool)
     },
 
-    async store(req, res){
-        console.log(req.body)
-        
-        const { title, link, description, tags } = req.body
-        if(!title || !link) return res.status(400).json({ error: "Campos obrigatorios nao foram informados."})
+    async store(req, res) {
+        //console.log(req.body)
 
-        const tool = await Tools.create({
-            title, link, description, tags
-        })
-        
-        return res.status(201).json(tool)
+        const { title, link, description, tags } = req.body
+        if (!title || !link) return res.status(400).json({ error: "Campos obrigatorios nao foram informados." })
+
+        try {
+            const tool = await Tools.create({
+                title, link, description, tags
+            })
+            return res.status(201).json(tool)
+        } catch (err) {
+            return res.status(400).send({ error: "Erro ao cadastrar uma nova ferramenta" })
+        }
+
     },
 
-    async delete(req,res){
+    async delete(req, res) {
         const { id } = req.params
-        if(!id) return res.status(400).json({ error: "ID nao informado"})
+
+        if(!mongoose.Types.ObjectId.isValid(id)){
+            return res.status(404).json({ error: "ID invalido" })
+        }
+        
+        if (!id) return res.status(400).json({ error: "ID nao informado" })
 
         await Tools.findByIdAndRemove(id)
